Add files via upload
To check whether any given number is an even or an odd number in C programming, you have to ask from the user to enter a number to check for even/odd using divisibility test by 2. That is if the given number is divisible by 2, then it will be an even number, otherwise it will be an odd number.